chore: upgrade platform-java to latest flatbuffers feed request ack
Changes:
- updated FBS generated classes
- add handler for FeedRequestAck message to example
- add feedId to market depth

+++ src/main/java/com/reactivemarkets/encoding/feed/FeedType.java
@@ -18,7 +18,7 @@ package com.reactivemarkets.encoding.feed;
 
 public final class FeedType {
   private FeedType() { }
-  public static final byte Default = 0;
+  public static final short Default = 0;
 
   public static final String[] names = { "Default", };
